requirements.txt
adan-pytorch==0.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.11.0
anykeystore==0.2
apex==0.9.10.dev0
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
commentjson==0.9.0
cryptacular==1.6.2
Cython==3.2.1
dateparser==1.2.2
defusedxml==0.7.1
fairseq==0.12.2
fastapi==0.121.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
h5py==3.15.1
hupper==1.12.1
hydra-core==1.0.7
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lark-parser==0.7.8
llvmlite==0.44.0
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.0.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
PasteDeploy==3.1.0
patsy==1.0.2
pbkdf2==1.3
pillow==12.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.5.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pyramid==2.0.2
pyramid-mailer==0.15.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
repoze.sendmail==4.4.1
requests==2.32.5
requests-oauthlib==2.0.0
sacrebleu==2.5.1
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.44.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
statsmodels==0.14.5
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.9.1
torchaudio==2.9.1
torchsde==0.2.6
torchvision==0.24.1
tqdm==4.67.1
trampoline==0.1.2
transaction==5.0
translationstring==1.4
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
velruse==1.1.1
venusian==3.1.1
wandb==0.23.0
WebOb==1.8.9
websocket-client==1.8.0
websockets==15.0.1
WTForms==3.2.1
wtforms-recaptcha==0.3.2
yarl==1.22.0
zope.deprecation==6.0
zope.interface==8.1.1
zope.sqlalchemy==4.0
